requirements.txt
asgiref==3.8.1
astroid==3.3.8
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
commit==0.1.0
dill==0.3.9
distlib==0.3.9
Django==5.1.4
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.44
identify==2.6.5
idna==3.10
isort==5.13.2
JustWatch==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pylint==3.3.3
pyproject-api==1.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
smmap==5.0.2
sqlparse==0.5.3
tomlkit==0.13.2
tox==4.23.2
urllib3==2.3.0
virtualenv==20.28.1
whatthepatch==1.0.7
